Replace the Runner class with an S3Export one
This is named to reflect that S3 is one of the approaches to export, it
will soon be complemented by File Export that will replace the CSV
splitter approach.

This S3 export essentially does the same thing as the previous Runner
task but leverages the new DailyReport class for reusability of common
processes.
